improved encoding speed using a tagged stream
This is a preliminary experiment, which currrently uses a preview release of
the new bytestring-builder. It uses an explicit datatype to represent the
stream of values to be encoded. This allows to get rid of PAP allocations and
unknown calls with respect to the existing approach.

My hypothesis for the not-so-good performance of the existing approach is that
it results in too many unknown, unsaturated calls due to the higher-order
nature of builders. This is a slow path of the unknown function call
implementation. Using an explicit type and a simple difference-list for
flattening results in THUNKs and saturated, unknown calls, which represents a
(as it seems, significantly) faster path.